refactor JSON:API findRecord op into multiple files
